backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gtts>=2.5.4",
    "openai>=1.70.0",
    "strawberry-graphql[asgi]>=0.263.0",
    "tinydb>=4.8.2",
    "uvicorn>=0.34.0",
]
